+++ highwind-meta-box-color.js
@@ -2,5 +2,6 @@
  * Highwind Light Color Picker
  */
 jQuery(document).ready(function($){
-	$('.highwind-header-color').wpColorPicker();
+	$('.highwind-header-bg-color').wpColorPicker();
+	$('.highwind-header-title-color').wpColorPicker();
 });