+++ changelog.md
@@ -1,7 +1,8 @@
 # Highwind Light changelog
 
 ## 1.1.1
-* ADDED   - Color picker for the Header. Now you can change the color of the header per post or page.
+* ADDED   - Color picker for the Header. Now you can change the color of the header and title per post or page.
+* ADDED   - Localization for United Kingdom (en_GB)
 * UPDATED - POT file.
 
 ## 1.0.1